+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## main
 
+* Create `CloseButton` component.
+
+    *Manuel Puyol*
+
 * **Breaking change**: Rename `AutoCompleteComponent` to `AutoComplete` and `AutoCompleteItemComponent` to `AutoComplete::Item`.
 
     *Manuel Puyol*