+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## main
 
-* Create `CloseButton` and `HiddenTextExpander` components.
+* Add `block` flag to `ButtonComponent`.
+
+    *Manuel Puyol*
+
+* Add `link` and `invisible` schemes to `ButtonComponent`.
+
+    *Manuel Puyol*
+
+* Create `CloseButton` and `HiddenTextExpander` component.
 
     *Manuel Puyol*